fixed a bug in raw.app, wrong data type for r, x, b, in line section

+++ include/nontransformerbranch.hpp
@@ -9,8 +9,8 @@ public:
     int i;
     int j;
     std::string ckt;
-    int r;
-    int x;
+    double r;
+    double x;
     double b;
     double ratea;
     double rateb;
